python_flask_user_crud_api/requirements.txt
quart==0.19.4
motor==3.5.1
bcrypt==4.2.0
pyjwt==2.9.0
python-multipart==0.0.9
aiofiles==23.2.1